+++ tests/src/test_memfault_trace_event.cpp
@@ -45,19 +45,22 @@ TEST(MfltTraceEvent, Test_CaptureButStorageUninitialized) {
   CHECK_EQUAL(rv, -1);
 }
 
-TEST(MfltTraceEvent, Test_CaptureOk) {
+
+static void prv_run_capture_test(void *pc, void *lr, size_t storage_size, size_t expected_encoded_size) {
   fake_memfault_event_storage_clear();
-  fake_memfault_event_storage_set_available_space(MEMFAULT_TRACE_EVENT_WORST_CASE_SIZE_BYTES);
-
-  CHECK_EQUAL(memfault_trace_event_boot(s_fake_event_storage_impl), 0);
-
+  fake_memfault_event_storage_set_available_space(storage_size);
   mock().expectOneCall("prv_begin_write");
-  mock().expectOneCall("prv_finish_write").withParameter("rollback", false);
-
-  void *pc = (void *)0x12345678;
-  void *lr = (void *)0xaabbccdd;
+  const bool expect_rollback = (storage_size < expected_encoded_size);
+  mock().expectOneCall("prv_finish_write").withParameter("rollback", expect_rollback);
   const int rv = memfault_trace_event_capture(pc, lr, kMfltTraceReasonUser_test);
-  CHECK_EQUAL(rv, 0);
+  CHECK_EQUAL(expect_rollback ? -2 : 0, rv);
+  mock().checkExpectations();
+}
+
+TEST(MfltTraceEvent, Test_CaptureOk_PcAndLr) {
+  fake_memfault_event_storage_clear();
+  const int rv = memfault_trace_event_boot(s_fake_event_storage_impl);
+  CHECK_EQUAL(0, rv);
 
   const uint8_t expected_data[] = {
       0xA7, 0x02, 0x02, 0x03, 0x01, 0x07, 0x69, 0x44,
@@ -68,6 +71,38 @@ TEST(MfltTraceEvent, Test_CaptureOk) {
       0x03, 0x02, 0x1A, 0x12, 0x34, 0x56, 0x78, 0x03,
       0x1A, 0xAA, 0xBB, 0xCC, 0xDD,
   };
+
+  // anything less than the expected size should fail to encode
+  for (size_t i = 1; i <= sizeof(expected_data); i++) {
+    void *pc = (void *)0x12345678;
+    void *lr = (void *)0xaabbccdd;
+    prv_run_capture_test(pc, lr, i, sizeof(expected_data));
+  }
+
+  fake_event_storage_assert_contents_match(expected_data, sizeof(expected_data));
+}
+
+TEST(MfltTraceEvent, Test_CaptureOk_LrOnly) {
+  fake_memfault_event_storage_clear();
+  const int rv = memfault_trace_event_boot(s_fake_event_storage_impl);
+  CHECK_EQUAL(0, rv);
+
+  const uint8_t expected_data[] = {
+      0xA7, 0x02, 0x02, 0x03, 0x01, 0x07, 0x69, 0x44,
+      0x41, 0x41, 0x42, 0x42, 0x43, 0x43, 0x44, 0x44,
+      0x0A, 0x64, 0x6D, 0x61, 0x69, 0x6E, 0x09, 0x65,
+      0x31, 0x2E, 0x32, 0x2E, 0x33, 0x06, 0x66, 0x65,
+      0x76, 0x74, 0x5F, 0x32, 0x34, 0x04, 0xA2, 0x06,
+      0x03, 0x03, 0x01,
+  };
+
+  // anything less than the expected size should fail to encode
+  for (size_t i = 1; i <= sizeof(expected_data); i++) {
+    void *pc = (void *)NULL;
+    void *lr = (void *)1;
+    prv_run_capture_test(pc, lr, i, sizeof(expected_data));
+  }
+
   fake_event_storage_assert_contents_match(expected_data, sizeof(expected_data));
 }
 
@@ -83,7 +118,7 @@ TEST(MfltTraceEvent, Test_CaptureStorageFull) {
   mock().expectOneCall("prv_finish_write").withParameter("rollback", true);
 
   const int rv = memfault_trace_event_capture(0, 0, kMfltTraceReasonUser_test);
-  CHECK_EQUAL(rv, -2);
+  CHECK_EQUAL(-2, rv);
 }
 
 TEST(MfltTraceEvent, Test_GetWorstCaseSerializeSize) {